### 2022-12-20 #187 @whedon
[PRE REVIEW]: Special relativity in R: the lorentz package
**Submitting author:** @RobinHankin (<a href="http://orcid.org/0000-0001-5982-0415">Robin Hankin</a>)
**Repository:** <a href="https://github.com/RobinHankin/lorentz" target ="_blank">https://github.com/RobinHankin/lorentz</a>
**Version:** 1.1-0
**Managing EiC:** Jordan Gorzalski

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode". You can read more about what that means in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss).
