### 2022-12-20 #187 @RobinHankin
Hello, not sure if a human reads this.  My suggestions for reviewers would be:

1.  Markus P\"ossel <poessel@hda-hd.de>    - influential worker in the field of  relativity education
2. Ian Hawke <I.Hawke@soton.ac.uk>  -  editor for a previous JOSE paper of mine which dealt with SR
