### 2022-12-30 #187 @labarba
We still do not have an assigned handling editor, but I wonder if we can entice Gabriele Bozzola, @Sbozzolo, to help us as a reviewer? (By complete coincidence, he just followed me on Fosstodon, ha ha! I followed back ☺️)
