### 2022-12-31 #187 @Sbozzolo
> Thanks for popping in here, @Sbozzolo ! Appreciate it. It looked like a lucky strike when I saw you are a numerical relativist. Do you know anyone who could help us with this review?

I'd be happy to help, but language is a barrier here. And, in general, I suspect it will be tough to find someone in the relativity and gravitational physics community, since we use primarily Python, Julia, C/C++, and Fortran. From a cursory look at this package, my impression is that it does not implement very sophisticated relativistic calculations, so you probably don't need a domain expert to follow what is going on (the paper itself claims that the package is meant to be accessible to undergraduate students).

One person that I see in the [list of volunteers for JOSE](https://docs.google.com/spreadsheets/d/1gQ_KZnJx-lyJwiuHBp9rE6b2ugMkJZzigvt4QLQHCH8/edit#gid=0) is HaoZeke, who lists R as one of their preferred languages and molecular dynamics simulations as one of their topics (which suggests a physics background).

From the [JOSS list](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822), I can see AlexAxthelm, which lists R and "physics, education" as their interests. This is the only partial match I found in the first 750 lines of the document, which gives you an idea of how niche is this combination of programming language and topic.
