### 2023-01-04 #187 @HaoZeke
Hi @arm61, happy to help out, but I'm a little backlogged, so I might get to this after the 15th at the earliest. If that's alright, I took a preliminary peek at the code-base and I will be able to handle the technical (`R`) review as well as the physics (at the level hinted at in the paper).
