### 2023-01-20 #187 @RobinHankin
Hello Andrew, I have to say that the intersection between relativistic
physics and R is almost nonexistent  [a situation I hope to improve with
this submission!].   You might ask Zurab Silagadze,  ***@***.***>
who I have been discussing the work with.   also Rachel E. Scherr <
***@***.***> and
Andrew Harris ***@***.***>
 who work in physics education and astronomy in R respectively.

Best wishes

Robin

--
Robin K. S. Hankin
Associate Professor, computational statistics

On Fri, Jan 20, 2023 at 9:54 PM Andrew McCluskey ***@***.***>
wrote:

> Still working on finding a second reviewer. @RobinHankin
> <https://github.com/RobinHankin> do you know of others working in physics
> with experience in R?
>
