### 2023-02-22 #196 @whedon
[REVIEW]: Special relativity in R: the lorentz package
**Submitting author:** @RobinHankin (<a href="http://orcid.org/0000-0001-5982-0415">Robin Hankin</a>)
**Repository:** <a href="https://github.com/RobinHankin/lorentz" target ="_blank">https://github.com/RobinHankin/lorentz</a>
**Version:** 1.1-0

- [ ] **Version:** Does the release version given match the GitHub release (1.1-0)?

- [ ] **Version:** Does the release version given match the GitHub release (1.1-0)?
