### 2023-02-27 #196 @arm61
@RobinHankin it looks like there some DOIs missing from the paper, it would be great if you can fill them in when you have a chance.
