### 2023-02-27 bd03475
further work on issue #43
