+++ inst/JOSE/ref.bib
@@ -5,7 +5,8 @@
   year = 	 {1997},
   volume =	 {27},
   number =	 {6},
-  pages =	 {881-951}
+  pages =	 {881-951},
+  doi =          {10.1007/bf02550347}
   }
 
 @Book{goldstein1980,
@@ -23,7 +24,8 @@
   year = 	 {1994},
   volume =	 {8},
   number =	 {1},
-  pages =	 {92--97}
+  pages =	 {92--97},
+  doi =          {10.1063/1.168517}
 }
 
 @Article{sherin2016,
@@ -35,7 +37,9 @@
   number =	 {5},
   pages =	 {369--374},
   month =        {May},
-  URL =          {https://github.com/MITGameLab/ OpenRelativity/}
+  URL =          {https://github.com/MITGameLab/ OpenRelativity/},
+  doi =          {10.1119/1.4938057}
+
   }
 
 
@@ -68,7 +72,7 @@
   doi = {10.48550/ARXIV.2212.07005},
   url = {https://arxiv.org/abs/2212.07005},
   author = {Hankin, Robin K. S.},
-  title = {A systematic search for a three-velocity gyrodistributive law in special relativity with the lorentz \proglang{R} package},
+  title = {A systematic search for a three-velocity gyrodistributive law in special relativity with the lorentz {R} package},
   publisher = {arXiv},
   year = {2022}
   }
@@ -80,7 +84,8 @@
   year = 	 {1905},
   volume =	 {322},
   number =	 {6},
-  pages =	 {132--148}
+  pages =	 {132--148},
+  doi =          {10.1002/andp.19053220607}
   }
 
 @Article{lorentz1904,
@@ -89,7 +94,8 @@
   journal = 	 {Proceedings of hte Royal Netherlands Academy of Arts and Sciences},
   year = 	 {1904},
   volume =	 {6},
-  pages =	 {809--831}
+  pages =	 {809--831},
+  doi =          {10.1007/978-94-015-3445-1_5}
   }
 
 @Article{einstein1907,
@@ -97,7 +103,8 @@
   title = 	 {Kosmologische Betrachtungen zur allgemeinen Relativit\"{a}tstheorie},
   journal = 	 {Sitzungsberichte der Preu{\ss }ischen Akademie der Wissenschaften},
   year = 	 {1907},
-  volume =	 {142}
+  volume =	 {142},
+  doi =          {10.1007/978-3-663-19510-8_9}
   }
 
 @misc{possel2018,
