### 2023-02-27 #196 @RobinHankin
Missing DOI issue fixed in [bd03475](https://github.com/RobinHankin/lorentz/commit/bd03475f05543ea4ff4e7a241f2ff1d792b1a0b2)

- Robin
