### 2023-04-08 #196 @IanHawke
Apologies for being slow about this - March was heavier than I expected. Will be back on this in about 10 days time. In my general look-through so far it has looked good, but I want to think about the examples and the role of Thomas precession (eg) in the docs a bit more.
