+++ README.md
@@ -11,7 +11,7 @@ The lorentz package: special relativity in R
 [![CRAN\_Status\_Badge](https://www.r-pkg.org/badges/version/lorentz)](https://cran.r-project.org/package=lorentz)
 [![Rdoc](https://www.rdocumentation.org/badges/version/lorentz)](https://www.rdocumentation.org/packages/lorentz)
 [![Codecov test
-coverage](https://codecov.io/gh/RobinHankin/lorentz/branch/master/graph/badge.svg)](https://codecov.io/gh/RobinHankin/lorentz/branch/master)
+coverage](https://app.codecov.io/gh/RobinHankin/lorentz/branch/master/graph/badge.svg)](https://app.codecov.io/gh/RobinHankin/lorentz/branch/master)
 <!-- badges: end -->
 
 # Overview
