+++ README.Rmd
@@ -22,7 +22,7 @@ knitr::opts_chunk$set(
 <!-- badges: start -->
 [![CRAN_Status_Badge](https://www.r-pkg.org/badges/version/lorentz)](https://cran.r-project.org/package=lorentz)
 [![Rdoc](https://www.rdocumentation.org/badges/version/lorentz)](https://www.rdocumentation.org/packages/lorentz)
-[![Codecov test coverage](https://codecov.io/gh/RobinHankin/lorentz/branch/master/graph/badge.svg)](https://codecov.io/gh/RobinHankin/lorentz/branch/master)
+[![Codecov test coverage](https://app.codecov.io/gh/RobinHankin/lorentz/branch/master/graph/badge.svg)](https://app.codecov.io/gh/RobinHankin/lorentz/branch/master)
 <!-- badges: end -->
 
 
