+++ NAMESPACE
@@ -5,7 +5,7 @@ importFrom("stats", "runif", "rnorm")
 importFrom("emulator","quad.tdiag","quad.tform", "quad.form","quad.3form", "quad.3tdiag")
 importFrom("graphics", "arrows", "legend", "plot", "points")
 importFrom("magic","adiag")
-import("magrittr")
+importFrom("magrittr", "%>%")
 
 importFrom("tensor","tensor")
 importFrom("utils","capture.output")
