+++ DESCRIPTION
@@ -1,11 +1,10 @@
 Package: lorentz
 Type: Package
 Title: The Lorentz Transform in Relativistic Physics
-Version: 1.1-0
+Version: 1.1-1
 Authors@R: person(given=c("Robin", "K. S."), family="Hankin", role = c("aut","cre"), email="hankin.robin@gmail.com", comment = c(ORCID = "0000-0001-5982-0415"))
-Depends: magrittr
 Suggests: knitr,testthat,rmarkdown
-Imports: emulator (>= 1.2-20),tensor,magic
+Imports: emulator (>= 1.2-20),tensor,magic,magrittr
 Maintainer: Robin K. S. Hankin <hankin.robin@gmail.com>
 Description: The Lorentz transform in special relativity; also the
     gyrogroup structure of three-velocities.  Performs active and
