### 2023-09-08 #196 @HaoZeke
Apologies @arm61 and @RobinHankin. I've started opening issues for the review, though it looks very good :)

I've opened issues with the specific concerns. Some small things left to finish (docs, apidocs, build systems, ci, contributing docs etc.), otherwise its all good. I didn't consider the celerity portions because they are marked as experimental.
