### 2023-11-13 #196 @HaoZeke
> @HaoZeke I see that @RobinHankin has closed most of the issues you opened are you able to now complete the review?

I'm not sure. @RobinHankin would it be possible to wrap up that issue sometime soon? If not,  I would still recommend acceptance (since the code is relatively simple, and there are tests), but it would be nicer to have it finished :)
