### 2023-11-13 #196 @RobinHankin
Hi there, sorry for the delay [I have had some family and personal issues
to deal with recently].

I've sorted out all the comments apart from the workflow comments and the
code coverage comments.
I spent a massive amount of time trying to sort out the automatic code
coverage and the travis checking, to no avail.  You can see from commits
such as

https://github.com/RobinHankin/lorentz/commit/dae66c15d416d03ebd23c9a572793dc418d613cc

that this was a frustrating and ultimately unsuccessful task.  As you say,
the code is relatively simple and indeed passes local tests [for me this is
"R CMD check"] and the test coverage is almost complete.  If Rohit
recommends acceptance then this would be a great relief from my mind.

With best wishes

Robin

--
Robin K. S. Hankin
Associate Professor, computational statistics

On Tue, Nov 14, 2023 at 9:30 AM Rohit Goswami ***@***.***>
wrote:

> @HaoZeke <https://github.com/HaoZeke> I see that @RobinHankin
> <https://github.com/RobinHankin> has closed most of the issues you opened
> are you able to now complete the review?
>
> I'm not sure. @RobinHankin <https://github.com/RobinHankin> would it be
> possible to wrap up that issue sometime soon? If not, I would still
> recommend acceptance (since the code is relatively simple, and there are
> tests), but it would be nicer to have it finished :)
>
