### 2024-04-04 #196 @arm61
@Hughshanahan, @jwagemann, or @dpshelio would any of you be interested in reviewing this JOSE paper?
