### 2024-08-28 #196 @arm61
Hey @jfoadi, if you comment

> @editorialbot generate my checklist

you should get a fresh review checklist to work through. More information about reviewing for JOSE can be found here (https://openjournals.readthedocs.io/en/jose/reviewer_guidelines.html) and if you have any questions please just ask.

Thanks again for agreeing to review for JOSE!
