### 2024-08-28 #196 @jfoadi
- [x] **Version:** Does the release version given match the GitHub release?
