### 2024-10-14 #196 @jfoadi
Hello Andrew, sorry for the late reply. I’m afraid the start of the semester here in Bath has been hell. I hope to see the light again at the end of October. I haven’t given up, don’t worry.
Best wishes,
James

Sent from Yahoo Mail for iPhone
