### 2025-03-10 #196 @arm61
@jfoadi just following up again to see if you can find the time to finish this off?
