### 2025-03-15 #196 @jfoadi
@RobinHankin I am now reading the article and will suggest improvements for readability, soundness, etc. First some quick fixes. Later some more fundamental critic.

1. I'd suggest the formal term "transformation/s" rather the informal "transform/s" throughout the text.
2. Line 5. Replace "supercede" with "supersede".
3. Lines 28 and 58. "the classical limit $c = \infty$" should be replaced by "the classical limit $c\ \rightarrow\ \infty$".
4. Line 34. "succintly offers", to be replaced by "succintly offered".
5. Line 45. Should this in line reference, https://github.com/nick1627/RelativityVisualisation, be added as standard reference?
6. Lines 20 and 49.
7. Line 56. "tensors of order two" should be written "tensors of order 2", as here we are in a technical context.
8. Line 57. Similarly, rather than "speed of light is one by default" you should write "the speed of light is 1 by default".
9. Line 59. Here you are suggesting a vignette from your package. It's a good opportunity to include the R code for running it. Many readers will be skilled with R, but some will not be.
10. "gyrogroup", "gyrodistributive law": these are terms specific of a specialised part of SR. Do you need to include them? Chances are readers have no heard of them. If you want to include them, at least spend a paragraph or two of explanation.
