### 2025-03-15 #196 @jfoadi
@RobinHankin I need this part of the vignette to be clarified, before moving on to the rest: " Suppose me and my friend have zero relative velocity, but my friend is in a spaceship and I am outside it, in free space, at rest ... Now his spaceship acquires a constant velocity, leaving me stationary" It;s a bit confusing. So, at the beginning, as you are at rest in free space, your friend too is at rest with respect to free space. Pedagogically, being in a spaceship makes you think you are moving, so you'll need to strengthen that. Can you confirm that at the beginning both you and your friend are at rest with respect to free space and then you remain at rest, but your friend moves with respect to it?
