### 2025-05-09 #196 @RobinHankin
Hello @arm61 and @jfoadi!  I have attended to all comments made bye @fjoadi at https://github.com/openjournals/jose-reviews/issues/196#issuecomment-2726497890 and https://github.com/openjournals/jose-reviews/issues/196#issuecomment-2726781919

Please see  https://github.com/RobinHankin/lorentz/issues/43 for an itemized link and the fixes to the issues raised.

Can you guys advise on the next step please?
