### 2025-05-09 #196 @jfoadi
3.	“@editorialbot generate pdf
