### 2025-05-09 #196 @jfoadi
@RobinHankin I understand the scenario with the spaceship now; thanks for changing and clarifying the text in the vignette. While I finish exploring the rest of the vignette, I have two comments:

1) I assume readers of "JOSE" already know from where packages must be downloaded. There is a "Repository" link on the left margin of the paper. In case this is not enough ( @arm61 any comment? ) you might want to add instructions on how to install the package, as it's not found anywhere in the article. Also, in your instructions for the vignette, you need to close the " you opened ("vignette("lorentz") ).

2) I tried installing the current version of lorentz from github and the vignette didn't work.

Give me another couple of days to finish trying out the vignette.
