### 2025-05-09 #196 @jfoadi
@RobinHankin I have actually completed the vignette faster than I thought. You might want to double-check some formatting (I don't need to look into that). But I think everything else is fine and I am happy for your article to be published.
