### 2025-05-12 #196 @jfoadi
@arm61 Most boxes of the checklist are ticked off. These ones remain unticked, but I'm not sure they need ticking: 1) Example Usage. The examples are not in the software article, but in the documentation/vignette; 2) I did not see tests explicitly written out, but they could be implicit as the package is in CRAN; 3) not sure about community guidelines.
