### 2025-05-12 #196 @RobinHankin
Hello @arm61 and @jfoadi

I can confirm that:

(1) *Example usage* is given in the vignette
(2) *tests* are given using the standard `testthat` mechanism [also the package passes the CRAN QC checks cleanly]
(3) *community guidelines* are given in the `CODE_OF_CONDUCT.md` and `CONTRIBUTING.md` files in the root directory on the repo

thanks!
