### 2025-05-13 #196 @arm61
Some quick things before we accept:
- Do you want to include a department in the affiliation, currently just reads "University of Stirling"?
- A few of the badges seem to be broken in the README.md
- Adding `{}` to get the right capitalisation in:
   - https://github.com/RobinHankin/lorentz/blob/87c6996033337d7a206a7ec8a03f525da7e21f38/inst/JOSE/ref.bib#L50
   - https://github.com/RobinHankin/lorentz/blob/87c6996033337d7a206a7ec8a03f525da7e21f38/inst/JOSE/ref.bib#L63
   - https://github.com/RobinHankin/lorentz/blob/87c6996033337d7a206a7ec8a03f525da7e21f38/inst/JOSE/ref.bib#L82
   - https://github.com/RobinHankin/lorentz/blob/87c6996033337d7a206a7ec8a03f525da7e21f38/inst/JOSE/ref.bib#L103
