### 2025-05-15 #196 @arm61
Can you now mint a DOI using Zenodo for the manuscript and make a versioned release of the repo? The Zenodo archive should have the same name as the paper.
