### 2025-05-15 #196 @RobinHankin
@arm61 thanks for this!

* uploaded to zenodo just now, they gave me `10.5281/zenodo/15421909`
* Tag `1.1-3` pushed to the repo just now.

Best wishes

Robin
