### 2025-05-15 #196 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
