### 2025-06-23 #196 @labarba
it looks like there was a misunderstanding, because I see the [Zenodo deposit](https://zenodo.org/records/15421909) has a copy of the JOSE paper PDF file. In fact, the Zenodo archive is meant to hold the package the paper is about (the full [repo](https://github.com/RobinHankin/lorentz)). Note that JOSE does make archival deposit of the published papers using the Portico service, as mentioned in our [Ethics guidelines](https://jose.theoj.org/about#ethics) section of the website.
