### 2025-06-24 #196 @RobinHankin
@labarba

thanks for this.   My response in **bold** below:

* Double check authors and affiliations (including ORCIDs) **confirmed**
* Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSE paper. **Git tag 1.1-5, corresponding to release 1.1-5, https://github.com/RobinHankin/lorentz/releases**
* Archive the release on Zenodo/figshare/etc and post the DOI here.  **https://zenodo.org/records/15421909**
* Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSE paper.  **checked**
* Make sure that the license listed for the archive is the same as the software license.  **check**
