### 2025-06-24 #196 @labarba
@RobinHankin – please see my [comment above](https://github.com/openjournals/jose-reviews/issues/196#issuecomment-2997423151) about the Zenodo archive
