### 2025-06-24 #196 @RobinHankin
@labarba  Sorry about this, I think I was getting confused.  I have uploaded the files of the repo to zenodo as instructed and have a DOI:

https://10.5281/zenodo.15731108

Best wishes, Robin
