### 2025-06-25 #196 @labarba
I noticed that the repository lists a GPL-3.0 license, while the Zenodo repository shows a CC-BY 4.0 license.

Perhaps you meant to apply a dual-licensing scheme, where code is under GPL but vignette content is under CC-BY?

You may add more than one license to a Zenodo archive when the files you are uploading are under different licenses; you can declare all applicable licenses in the Zenodo record metadata.
https://help.zenodo.org/docs/deposit/describe-records/licenses/#:~:text=Mixed%20license%20uploads,done%20in%20the%20example%20below:
