### 2025-06-26 #196 @RobinHankin
@labarba   thanks for this,  I did not know one could have dual-licensing.  I have added GPL to the licenses on Zenodo.  Best wishes, Robin
