### 2025-06-26 #196 @labarba
I think it's important because GPL is copyleft, whereas CC-BY is not
