### 2025-06-27 #196 @labarba
Interesting. I now see only GPL, rather than both. No matter, I will go ahead and publish as at least now the licenses match between repo and archive.
