### 2025-06-27 #196 @RobinHankin
looks good to me except two very minor typos.

Line 30 of page 1, a  superfluous "in".

```
Old text: "familiar matrix multiplication, and in addition incorporates in an efficient and consistent"
New text: "familiar matrix multiplication, and in addition incorporates an efficient and consistent"
```

Line 32 of page 1, "mometum" should be "momentum":

```
Old text: passive transformations, four-mometum of photons, and in particular the difficult and
New text: passive transformations, four-momentum of photons, and in particular the difficult and
```

I have updated the repo to reflect this.

Best wishes

Robin
