### 2025-06-28 #196 @labarba
@openjournals/dev – could you take a quick look here at this recently published paper? The title shows visible `<code></code>` tags and presumably will be deposited in Crossref like that, which is minor but undesirable...
