### 2025-07-01 #196 @RobinHankin
@arm61  thanks for this, I appreciate it!  I actually supervised a student using the software not long ago,  it is working as intended!
